Bug 1100499 - Add GMP API to enumerate records stored on disk. r=jesup
UltraBlame original commit: 405b0c00071f7f583f722a48e6537950e1a29653

+++ dom/media/gmp/gmp-api/gmp-audio-codec.h
@@ -37,7 +37,7 @@ struct GMPAudioCodec
   
   
   const uint8_t* mExtraData;
-  size_t         mExtraDataLen;
+  uint32_t       mExtraDataLen;
 };
 
 #endif
